auto infer appState
