+++ src/window_main/components/collection/CollectionTable.tsx
@@ -48,7 +48,7 @@ import {
   CollectionTableProps
 } from "./types";
 import { useSelector } from "react-redux";
-import { AppState } from "../../../shared/redux/appState";
+import { AppState } from "../../../shared/redux/reducers";
 
 function isBoosterMathValid(filters: Filters<CardsData>): boolean {
   let hasCorrectBoosterFilter = false;
