+++ src/window_main/components/misc/ShareButton.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { AppState } from "../../../shared/redux/appState";
+import { AppState } from "../../../shared/redux/reducers";
 import { useSelector, useDispatch } from "react-redux";
 import { rendererSlice } from "../../../shared/redux/reducers";
