+++ src/window_main/components/popups/Share.tsx
@@ -2,7 +2,7 @@ import React, { useState, useCallback, useEffect } from "react";
 import ReactSelect from "../../../shared/ReactSelect";
 import { ipcSend } from "../../rendererUtil";
 import { useSelector, useDispatch } from "react-redux";
-import { AppState } from "../../../shared/redux/appState";
+import { AppState } from "../../../shared/redux/reducers";
 import { rendererSlice } from "../../../shared/redux/reducers";
 
 function shareTypeId(type: string): number {
