+++ src/window_main/components/settings/SectionData.tsx
@@ -10,7 +10,7 @@ import ReactSelect from "../../../shared/ReactSelect";
 import { parse, isValid } from "date-fns";
 import Button from "../misc/Button";
 import { useSelector } from "react-redux";
-import { AppState } from "../../../shared/redux/appState";
+import { AppState } from "../../../shared/redux/reducers";
 
 const LANGUAGES = [
   "en",
